web/requirements.txt
Django==1.10.5
gunicorn==19.3.0
psycopg2==2.6

django-dotenv==1.3.0

django-debug-toolbar==1.4
sqlparse==0.1.18

dealer==2.0.5
django-configurations==2.0
django-registration-redux==1.4
django-two-factor-auth==1.5.0
django-widget-tweaks==1.4.1
django-braces==1.11.0
Pillow==4.0.0
djangorestframework==3.5.4
djangorestframework-gis==0.11
